BLIT: only add impulse from ring buffer when needed
This'll allow to greatly diminish the size of the ring buffer: For the
moment, to generate a wave without artefacts, RB should be bigger than
the size of a full oscillation (which already made it too small for
anything below ~44hz). Now, it can be the size of the impulse.